challenges/forensics/static/requirements.txt
pillow==9.2.0
numpy==1.23.3
